Fix params validation + specs
* Move params validation to corresponding class: We only need to valide the
params when they need to be set / are actually used. Since Puppet 4.6 or 4.7,
stdlib's validate_string function doesn't accept undef anymore, so this has
to be changed.

* Remove specific error checking in specs: types checking validate_* functions
have different error messages depending on the version the stdlib module, the
version of puppet, maybe the version of ruby, the phase of the Moon... who knows.